APB-7591[CS] update response schema + split Des/If connector
