+++ test/uk/gov/hmrc/agentclientauthorisation/service/PostcodeServiceSpec.scala
@@ -33,20 +33,20 @@ import scala.concurrent.Future
 import uk.gov.hmrc.http.HeaderCarrier
 
 class PostcodeServiceSpec extends UnitSpec with MockitoSugar with BeforeAndAfterEach with TransitionInvitation {
-  val desConnector: DesConnector = mock[DesConnector]
+  val ifConnector: IfConnector = mock[IfConnector]
 
-  val service = new PostcodeService(desConnector)
+  val service = new PostcodeService(ifConnector)
 
   implicit val hc = HeaderCarrier()
 
   override protected def beforeEach(): Unit = {
     super.beforeEach()
-    reset(desConnector)
+    reset(ifConnector)
   }
 
   "clientPostcodeMatches" should {
     "return no error status if there is a business partner record with a matching UK postcode" in {
-      when(desConnector.getBusinessDetails(nino1)).thenReturn(
+      when(ifConnector.getBusinessDetails(nino1)).thenReturn(
         Future successful Some(BusinessDetails(Seq(BusinessData(Some(BusinessAddressDetails("GB", Some("AA11AA"))))), Some(MtdItId("mtdItId")))))
 
       val maybeResult = await(service.postCodeMatches(nino1.value, "AA11AA"))
@@ -54,7 +54,7 @@ class PostcodeServiceSpec extends UnitSpec with MockitoSugar with BeforeAndAfter
     }
 
     "fail with a 501 if there is a business partner record with a matching non-UK postcode" in {
-      when(desConnector.getBusinessDetails(nino1)).thenReturn(
+      when(ifConnector.getBusinessDetails(nino1)).thenReturn(
         Future successful Some(BusinessDetails(Seq(BusinessData(Some(BusinessAddressDetails("US", Some("AA11AA"))))), Some(MtdItId("mtdItId")))))
 
       await(service.postCodeMatches(nino1.value, "AA11AA").failed) match {
@@ -63,7 +63,7 @@ class PostcodeServiceSpec extends UnitSpec with MockitoSugar with BeforeAndAfter
     }
 
     "fail with a 403 if there is a business partner record with a mis-matched postcode" in {
-      when(desConnector.getBusinessDetails(nino1)).thenReturn(
+      when(ifConnector.getBusinessDetails(nino1)).thenReturn(
         Future successful Some(BusinessDetails(Seq(BusinessData(Some(BusinessAddressDetails("GB", Some("ZZ99ZZ"))))), Some(MtdItId("mtdItId")))))
       await(service.postCodeMatches(nino1.value, "AA11AA").failed) match {
         case FailedResultException(r) => r.header.status shouldBe 403
@@ -71,7 +71,7 @@ class PostcodeServiceSpec extends UnitSpec with MockitoSugar with BeforeAndAfter
     }
 
     "return 403 if there is a business partner record with a no postcode" in {
-      when(desConnector.getBusinessDetails(nino1))
+      when(ifConnector.getBusinessDetails(nino1))
         .thenReturn(Future successful Some(BusinessDetails(Seq(BusinessData(Some(BusinessAddressDetails("GB", None)))), Some(MtdItId("mtdItId")))))
       await(service.postCodeMatches(nino1.value, "AA11AA").failed) match {
         case FailedResultException(r) => r.header.status shouldBe 403
@@ -79,14 +79,14 @@ class PostcodeServiceSpec extends UnitSpec with MockitoSugar with BeforeAndAfter
     }
 
     "return 403 if no business partner record is found" in {
-      when(desConnector.getBusinessDetails(nino1)).thenReturn(Future successful None)
+      when(ifConnector.getBusinessDetails(nino1)).thenReturn(Future successful None)
       await(service.postCodeMatches(nino1.value, "AA11AA").failed) match {
         case FailedResultException(r) => r.header.status shouldBe 403
       }
     }
 
     "return 403 if a business partner record is returned with no business data records" in {
-      when(desConnector.getBusinessDetails(nino1))
+      when(ifConnector.getBusinessDetails(nino1))
         .thenReturn(Future successful Some(BusinessDetails(Seq(), Some(MtdItId("mtdItId")))))
       await(service.postCodeMatches(nino1.value, "AA11AA").failed) match {
         case FailedResultException(r) => r.header.status shouldBe 403
@@ -94,7 +94,7 @@ class PostcodeServiceSpec extends UnitSpec with MockitoSugar with BeforeAndAfter
     }
 
     "return 403 if a business partner record is returned with multiple business data records" in {
-      when(desConnector.getBusinessDetails(nino1)).thenReturn(Future successful Some(BusinessDetails(
+      when(ifConnector.getBusinessDetails(nino1)).thenReturn(Future successful Some(BusinessDetails(
         Array(BusinessData(Some(BusinessAddressDetails("GB", Some("AA11AA")))), BusinessData(Some(BusinessAddressDetails("GB", Some("AA11AA"))))),
         Some(MtdItId("mtdItId"))
       )))
